Updated endpoint for delete

+++ src/main/kotlin/no/uio/bedreflyt/api/controller/triplestore/FloorController.kt
@@ -125,7 +125,10 @@ class FloorController (
     @DeleteMapping("/{floorNumber}")
     fun deleteFloor(@ApiParam(value = "Floor number", required = true) @PathVariable floorNumber: Int) : ResponseEntity<String> {
         log.info("Deleting floor $floorNumber")
-
+        
+        if (floorService.getFloorByNumber(floorNumber) == null) {
+            return ResponseEntity.notFound().build()
+        }
         if (!floorService.deleteFloor(floorNumber)) {
             return ResponseEntity.notFound().build()
         }
